bin/find_m4.sh: fix checksum logic in populate_db
I think I meant to write it with 'if ! ...' originally, because if we
exit 1, it means changecom is there, and stripped_contents is both
useless (empty) and we can't use it sanely.

Rewrite it such that we only use its value (as stripped contents) for
checksumming if we get 0 as an exit status.

Signed-off-by: Sam James <sam@gentoo.org>