bin/find_m4.sh: initial "unknown macro" database
Store unknown/new macros we encounter during scanning (not indexing) so
that we don't spam lots of output about "new" macros which are actually
the same thing we're seeing across many directories.

Such new macros are likely candidates indicating a missing repo or source
of macros we should be indexing.

Signed-off-by: Sam James <sam@gentoo.org>