bin/populate_m4_db.sh: fix importing no-serial macros
Fix importing no-serial macros where no other macros where
changed in the same commit which *did* have a serial number.

In that case, we'd end up with temp= being clobbered and nothing
putting the old temp= in batch_dirs.

(Note that we still have the optimization for not repeatedly
grepping; I think this was really just a leftover from when
we only wanted serials and I didn't go back and fix it. I should
have at least commented the assumption when I first wrote it...)

Signed-off-by: Sam James <sam@gentoo.org>